requirements.txt
streamlit

numpy

Pillow

requests
json
base64
